requirements.txt
pysqlite3-binary>=0.5.2
streamlit>=1.31.0
altair==5.4.1
anthropic==0.18.1
chromadb==0.4.24
python-dotenv==1.0.1
scikit-learn>=1.4.0
pydantic==1.10.12
numpy>=1.24.3,<2.0.0
packaging>=16.8,<25.0
certifi>=2024.6.2,<2024.8.0
charset-normalizer==3.4.0
click==8.1.7
gitdb==4.0.11
gitpython==3.1.43
idna==3.4
jsonschema==4.23.0
markdown-it-py==3.0.0
markupsafe==2.1.3
pillow>=9.5.0
protobuf==3.20.3
pyarrow>=11.0.0
pygments==2.18.0
python-dateutil==2.8.2
pytz>=2023.3,<2025
requests==2.31.0
six==1.16.0
tornado>=6.3.3
typing-extensions>=4.10.0,<5
urllib3>=2.0.0,<2.2
validators==0.20.0
setuptools>=60.0.0
